+++ example/lib/main.dart
@@ -35,7 +35,7 @@ class _MyHomePageState extends State<MyHomePage> {
   List<File> files = [];
 
   void _pickImage() async {
-    files.addAll(await AdvImagePicker.pickImagesToFile(context, maxSize: 250));
+    files.addAll(await AdvImagePicker.pickImagesToFile(context, maxSize: 1250));
 
     setState(() {});
   }